Update README.md
This project is of the game Mastermind. It is implemented using multiple things learnt in class; for loops, while loops, if/else statements, inputs and fstrings, as well using the external random library. The scope of the project is to implement the code breaking game in a simple way, breaking it up into functions that can be easily called upon and used throughout the game, as well as learning how to make a game in python.


The code is split up into different functions for every part of the game as follows:
1. generate_code generates the random code for the game in the range of 1-6. This is small enough to be playable while also being challenging for the player. 
2. get_guess is a user input and it evaluates whether the input is valid, filtering our invalid inputs and returning an error.
3. analyze_guess compares the user input to the random string and checks to see if it is in the correct position and if it is the correct number.
4. display_grid displays the grid of the numbers displayed, where it goes first number, second, third, fourth.
5. play_game is the main function that controls the main game as well as keeping track of the amount of attempts, storing the guesses, as well as calling the display function.